+++ src/components/FormikForm/FormikForm.tsx
@@ -54,6 +54,8 @@ import {
   ExpressionInputProps as ExpressionInputLocalProps,
   ExpressionInput as ExpressionInputLocal
 } from '../ExpressionInput/ExpressionInput'
+import { DataTooltipInterface } from '../../frameworks/Tooltip/types'
+import { HarnessDocTooltip } from '../../frameworks/Tooltip/Tooltip'
 
 const IsOptionLabel = '(optional)'
 const isObject = (obj: any): boolean => obj !== null && typeof obj === 'object'
@@ -76,6 +78,7 @@ export interface FormikContextProps<T> {
   optionalLabel?: string
   isOptional?: boolean // default to false
   autoComplete?: string
+  tooltipProps?: DataTooltipInterface
 }
 
 export interface TagInputProps<T> extends Omit<IFormGroupProps, 'labelFor' | 'items'> {
@@ -102,13 +105,17 @@ function TagInput<T>(props: TagInputProps<T> & FormikContextProps<any>) {
     inline = formik?.inline,
     tagInputProps,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
       labelFor={name}
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -156,6 +163,7 @@ function KVTagInput(props: KVTagInputProps & FormikContextProps<any>) {
     tagsProps,
     isOptional = false,
     optionalLabel = IsOptionLabel,
+    tooltipProps,
     ...restProps
   } = props
   const hasError = errorCheck(name, formik)
@@ -174,10 +182,14 @@ function KVTagInput(props: KVTagInputProps & FormikContextProps<any>) {
     return () => unregister(mentionsType)
   }, [])
 
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
+
   return (
     <FormGroup
       labelFor={name}
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -229,6 +241,7 @@ function MultiInput(props: MultiInputProps & FormikContextProps<any>) {
     tagsProps,
     optionalLabel = IsOptionLabel,
     isOptional = false,
+    tooltipProps,
     ...restProps
   } = props
   const hasError = errorCheck(name, formik)
@@ -247,9 +260,13 @@ function MultiInput(props: MultiInputProps & FormikContextProps<any>) {
     return () => unregister(mentionsType)
   }, [])
 
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
+
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -292,12 +309,16 @@ const CustomRender = (props: CustomRenderProps & FormikContextProps<any>) => {
     inline = formik?.inline,
     label,
     render,
+    tooltipProps,
     ...rest
   } = restProps
 
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -335,11 +356,16 @@ const FileInput = (props: FileInputProps & FormikContextProps<any>) => {
     onChange,
     inputProps,
     multiple = false,
+    tooltipProps,
     ...rest
   } = restProps
+
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -390,6 +416,7 @@ const RadioGroup = (props: RadioGroupProps & FormikContextProps<any>) => {
     label,
     radioGroup,
     onChange,
+    tooltipProps,
     ...rest
   } = props
 
@@ -400,9 +427,12 @@ const RadioGroup = (props: RadioGroupProps & FormikContextProps<any>) => {
       className: cx(cssRadio.radio, className)
     }
   })
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -430,7 +460,7 @@ export interface CheckboxProps extends UiKitCheckboxProps, Omit<IFormGroupProps,
 }
 
 const CheckBox = (props: CheckboxProps & FormikContextProps<any>) => {
-  const { formik, name, label, ...restProps } = props
+  const { formik, name, label, isOptional, optionalLabel, ...restProps } = props
   const hasError = errorCheck(name, formik)
   const {
     intent = hasError ? Intent.DANGER : Intent.NONE,
@@ -439,15 +469,29 @@ const CheckBox = (props: CheckboxProps & FormikContextProps<any>) => {
     inline = formik?.inline,
     onChange,
     className = '',
+    tooltipProps,
     ...rest
   } = restProps
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup labelFor={name} helperText={helperText} intent={intent} disabled={disabled} {...rest}>
       <UiKitCheckbox
         {...rest}
         className={cx(checkBoxCss.checkbox, className)}
         name={name}
-        label={label}
+        // eslint-disable-next-line
+        // @ts-ignore
+        label={
+          labelText ? (
+            <HarnessDocTooltip
+              className={css.checkBoxDocTooltipLabel}
+              tooltipId={tooltipProps?.dataTooltipId}
+              labelText={labelText}
+            />
+          ) : (
+            labelText
+          )
+        }
         inline={inline}
         disabled={disabled}
         checked={get(formik?.values, name)}
@@ -483,14 +527,18 @@ const MultiSelect = (props: MultiSelectProps & FormikContextProps<any>) => {
     placeholder,
     multiSelectProps,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
   const formikValue = get(formik?.values, name, [])
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -549,13 +597,17 @@ const Select = (props: SelectProps & FormikContextProps<any>) => {
     selectProps,
     onChange,
     value,
+    tooltipProps,
     ...rest
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -605,12 +657,16 @@ const Text = (props: TextProps & FormikContextProps<any>) => {
     label,
     placeholder,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -660,6 +716,7 @@ const ExpressionInput = (props: ExpressionInputProps & FormikContextProps<any>) 
     onChange,
     isOptional,
     optionalLabel = IsOptionLabel,
+    tooltipProps,
     ...restProps
   } = props
   const hasError = errorCheck(name, formik)
@@ -673,11 +730,14 @@ const ExpressionInput = (props: ExpressionInputProps & FormikContextProps<any>) 
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
 
   return (
     <FormGroup
       labelFor={name}
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       helperText={helperText}
       intent={intent}
       disabled={disabled}
@@ -718,14 +778,17 @@ const TextArea = (props: TextAreaProps & FormikContextProps<any>) => {
     label,
     textArea,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -908,6 +971,7 @@ const FormMultiTypeInput = (props: FormMultiTypeInputProps & FormikContextProps<
     helperText = hasError ? get(formik?.errors, name) : null,
     disabled = formik?.disabled,
     label,
+    tooltipProps,
     ...rest
   } = restProps
   const onChangeCallback: MultiTypeInputProps['onChange'] = useCallback(
@@ -919,9 +983,12 @@ const FormMultiTypeInput = (props: FormMultiTypeInputProps & FormikContextProps<
     [formik, multiTypeInputProps]
   )
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -968,6 +1035,7 @@ const FormMultiSelectTypeInput = (props: FormMultiSelectTypeInputProps & FormikC
     multiSelectTypeInputProps,
     isOptional = false,
     optionalLabel = IsOptionLabel,
+    tooltipProps,
     ...restProps
   } = props
   const hasError = errorCheck(name, formik)
@@ -979,9 +1047,12 @@ const FormMultiSelectTypeInput = (props: FormMultiSelectTypeInputProps & FormikC
     ...rest
   } = restProps
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1042,6 +1113,7 @@ const FormMultiTextTypeInput = (props: FormMultiTextTypeInputProps & FormikConte
     helperText = hasError ? get(formik?.errors, name) : null,
     disabled = formik?.disabled,
     label,
+    tooltipProps,
     ...rest
   } = restProps
   const value = get(formik?.values, name, '')
@@ -1057,10 +1129,12 @@ const FormMultiTextTypeInput = (props: FormMultiTextTypeInputProps & FormikConte
   )
 
   const autoComplete = props.autoComplete || getDefaultAutoCompleteValue()
-
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1103,14 +1177,17 @@ const FormCategorizedSelect = (props: FormCategorizedSelect & FormikContextProps
     inline = formik?.inline,
     categorizedSelectProps,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
   const value = get(formik?.values, name)
-
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1165,14 +1242,17 @@ const FormSelectWithSubview = (props: FormSelectWithSubviewProps & FormikContext
     changeViewButtonLabel,
     subview,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
   const value = get(formik?.values, name)
-
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}
@@ -1231,14 +1311,17 @@ const FormMultiSelectWithSubview = (props: FormMultiSelectWithSubviewProps & For
     changeViewButtonLabel,
     subview,
     onChange,
+    tooltipProps,
     ...rest
   } = restProps
 
   const value: MultiSelectOption[] = get(formik?.values, name)
-
+  const labelText = !isOptional ? label : `${label} ${optionalLabel}`
   return (
     <FormGroup
-      label={!isOptional ? label : `${label} ${optionalLabel}`}
+      label={
+        labelText ? <HarnessDocTooltip tooltipId={tooltipProps?.dataTooltipId} labelText={labelText} /> : labelText
+      }
       labelFor={name}
       helperText={helperText}
       intent={intent}