+++ src/index.ts
@@ -91,7 +91,6 @@ export { SparkChart } from './components/SparkChart/SparkChart'
 export { useModalHook, ModalProvider, ModalContext } from './components/Modal'
 export { ModalExample } from './components/ModalExample/ModalExample'
 export { timeSeriesChartDefaultConfig } from './components/TimeSeriesChart/TimeSeriesChart'
-export { MultiLogsExample } from './components/MultiLogsExample/MultiLogsExample'
 export { columnChartDefaultConfig } from './components/ColumnChart/ColumnChart'
 export { Table, TableProps } from './components/Table/Table'
 export { FormikForm, Formik, FormInput } from './components/FormikForm/FormikForm'
@@ -133,10 +132,18 @@ export {
   MultiSelectWithSubview,
   MultiSelectWithSubviewProps
 } from './components/MultiSelectWithSubView/MultiSelectWithSubView'
-export { MultiLogsViewer, MultiLogsViewerProps } from './components/MultiLogsViewer/MultiLogsViewer'
 export {
   WeightedStack,
   WeightedStackProps,
   WeightedStackData,
   LabelPosition
 } from './components/WeightedStack/WeightedStack'
+export { HarnessDocTooltip, useTooltips } from './frameworks/Tooltip/Tooltip'
+export { TooltipContext, TooltipContextProvider, useTooltipContext } from './frameworks/Tooltip/TooltipContext'
+export {
+  TooltipContextProviderProps,
+  TooltipContextValue,
+  TooltipRenderProps,
+  UseTooltipsReturn,
+  DataTooltipInterface
+} from './frameworks/Tooltip/types'