+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.154](https://github.com/wings-software/uicore/compare/v0.1.153...v0.1.154) (2021-05-10)
+
+
+### Features
+
+* [Tooltips]: Enabling tooltips for text and button components, and adding className prop ([#201](https://github.com/wings-software/uicore/issues/201)) ([193ac5f](https://github.com/wings-software/uicore/commit/193ac5f241ae85fe2a72085f24ff20be97a5de4a))
+
+### [0.1.153](https://github.com/wings-software/uicore/compare/v0.1.152...v0.1.153) (2021-05-10)
+
+
+### Features
+
+* [CDNG-7579]: Added new icons ([#203](https://github.com/wings-software/uicore/issues/203)) ([0ba5947](https://github.com/wings-software/uicore/commit/0ba594715ae54ae2bcc488b6af78b4ef098533b6))
+
+### [0.1.152](https://github.com/wings-software/uicore/compare/v0.1.151...v0.1.152) (2021-05-10)
+
+
+### Features
+
+* Update React version to 17 ([#169](https://github.com/wings-software/uicore/issues/169)) ([2039ee9](https://github.com/wings-software/uicore/commit/2039ee928d3fedaf5e29f632fefd548244d067a5))
+
+### [0.1.151](https://github.com/wings-software/uicore/compare/v0.1.150...v0.1.151) (2021-05-10)
+
+
+### Others
+
+* remove old logs code ([#200](https://github.com/wings-software/uicore/issues/200)) ([d5c3ffc](https://github.com/wings-software/uicore/commit/d5c3ffc5cc8fc508791090b1176a86c9b85cf15b))
+
+### [0.1.150](https://github.com/wings-software/uicore/compare/v0.1.149...v0.1.150) (2021-05-08)
+
+
+### Features
+
+* [Tooltips]: Tooltip component and corresponding context ([#162](https://github.com/wings-software/uicore/issues/162)) ([#198](https://github.com/wings-software/uicore/issues/198)) ([e1803a2](https://github.com/wings-software/uicore/commit/e1803a286034253ae0d380ed735de812ce238a33))
+
 ### [0.1.149](https://github.com/wings-software/uicore/compare/v0.1.148...v0.1.149) (2021-05-06)
 
 ### [0.1.148](https://github.com/wings-software/uicore/compare/v0.1.147...v0.1.148) (2021-05-06)